// app/utils/types/account.ts
export interface AccountDetails {
    avatar: Avatar;
    id: number;
    iso_639_1: string;
    iso_3166_1: string;
    name: string;
    include_adult: boolean;
    username: string;
}

// --- Avatar Types ---
export interface Avatar {
    gravatar: Gravatar;
    tmdb: Tmdb;
}

export interface Gravatar {
    hash: string;
}

export interface Tmdb {
    avatar_path: null;
}